+++ app/controllers/parameter_handler.py
@@ -1,10 +1,9 @@
 import app.models.models as models
 from app.models.response_models import ResponseModel
 import app.mysql.models as mysql_models
-from app.mysql.mysql import Nexus1DataBase
+from sqlalchemy.orm import Session
 
 import app.utils.vars as var
-from sqlalchemy.orm import Session
 
 
 class Parameter_Controller:
@@ -12,17 +11,16 @@ class Parameter_Controller:
     Controller for managing Parameter-related operations.
 
     This class handles the creation, updating, deletion, and retrieval of Parameters in the database.
-
-    
     """
 
-    def __init__(self) -> None:
+    def __init__(self, db: object) -> None:
         """
         Initializes a new instance of the Parameter_Controller class.
 
-        This constructor does not take any parameters and does not perform any operation.
+        Parameters:
+            db (object): An instance of a database (e.g., Nexus1DataBase) that will be used to create the database connection.
         """
-        pass
+        self.db = db  # Recibe la instancia de la base de datos desde fuera
     
     def healthz(self):
         """
@@ -43,26 +41,25 @@ class Parameter_Controller:
         and saves it to the database.
 
         Parameters:
-            body (models.Parameter): An object containing the shelter data to create.
+            body (models.ParameterCreate): An object containing the parameter data to create.
 
         Returns:
             ResponseModel: A response model with the status of the operation, message, and response data.
         """
         try:
-            body_row = mysql_models.Parameter(name = body.name, description= body.description)
-            db = Nexus1DataBase(var.MYSQL_URL)
-            with Session(db.engine) as session:
+            body_row = mysql_models.Parameter(name=body.name, description=body.description)
+            with Session(self.db.engine) as session:
                 session.add(body_row)
                 session.commit()
-                session.close()
+                session.refresh(body_row)
             return ResponseModel(
                 status="ok",
                 message="Parameter inserted into database successfully",
-                data=None,
+                data=body_row,
                 code=201
             )
         except Exception as e:
-            print("Error inserting Parameter into database")
+            print(f"Error inserting Parameter into database: {e}")
             return ResponseModel(
                 status="error",
                 message=str(e),
@@ -77,22 +74,20 @@ class Parameter_Controller:
         This method queries all Parameters records in the database and returns them.
 
         Returns:
-            ResponseModel: A response model with the status of the operation, message, and Parameter data.
+            ResponseModel: A response model with the status of the operation, message, and parameter data.
         """
         try:
-            db = Nexus1DataBase(var.MYSQL_URL)
             response: list = []
-            with Session(db.engine) as session:
+            with Session(self.db.engine) as session:
                 response = session.query(mysql_models.Parameter).all()
-                session.close()
             return ResponseModel(
                 status="ok",
                 message="All Parameters successfully retrieved",
                 data=response,
-                code=201
+                code=200
             )
         except Exception as e:
-            print("Error retrieving Parameters from database")
+            print(f"Error retrieving Parameters from database: {e}")
             return ResponseModel(
                 status="error",
                 message=str(e),
@@ -104,29 +99,35 @@ class Parameter_Controller:
         """
         Deletes a Parameter from the database.
 
-        This method takes a Parameters object, which contains the ID of the Parameters to delete.
+        This method takes a Parameter object, which contains the ID of the Parameter to delete.
 
         Parameters:
-            body (models.ParametersDelete): An object containing the shelter ID to delete.
+            body (models.ParameterDelete): An object containing the parameter ID to delete.
 
         Returns:
-            ResponseModel: A response model with the status of the operation, message, and deleted Parameters data.
+            ResponseModel: A response model with the status of the operation, message, and deleted parameter data.
         """
         try:
-            db = Nexus1DataBase(var.MYSQL_URL)
-            with Session(db.engine) as session:
+            with Session(self.db.engine) as session:
                 parameter_deleted = session.query(mysql_models.Parameter).get(body.id)
-                session.delete(parameter_deleted)
-                session.commit()
-                session.close()
-            return ResponseModel(
-                status="ok",
-                message="Parameter successfully deleted",
-                data=parameter_deleted,
-                code=201
-            )
+                if parameter_deleted:
+                    session.delete(parameter_deleted)
+                    session.commit()
+                    return ResponseModel(
+                        status="ok",
+                        message="Parameter successfully deleted",
+                        data=parameter_deleted,
+                        code=200
+                    )
+                else:
+                    return ResponseModel(
+                        status="error",
+                        message="Parameter not found",
+                        data=None,
+                        code=404
+                    )
         except Exception as e:
-            print("Error deleting Parameter from database")
+            print(f"Error deleting Parameter from database: {e}")
             return ResponseModel(
                 status="error",
                 message=str(e),
@@ -142,28 +143,33 @@ class Parameter_Controller:
         and updates the corresponding record in the database.
 
         Parameters:
-            body (models.ParametersUpdate): An object containing the updated Parameters data.
+            body (models.ParameterUpdate): An object containing the updated parameter data.
 
         Returns:
-            ResponseModel: A response model with the status of the operation, message, and updated Parameter.
+            ResponseModel: A response model with the status of the operation, message, and updated parameter.
         """
         try:
-            db = Nexus1DataBase(var.MYSQL_URL)
-            with Session(db.engine) as session:
-                Parameter: mysql_models.Parameter = session.query(mysql_models.Parameter).get(body.id)
-                Parameter.name = body.name
-                Parameter.description = body.description
-                session.dirty  # This seems redundant; the session will be dirty when an attribute is modified
-                session.commit()
-                session.close()
-            return ResponseModel(
-                status="ok",
-                message="Parameter successfully updated",
-                data=Parameter,
-                code=201
-            )
+            with Session(self.db.engine) as session:
+                parameter: mysql_models.Parameter = session.query(mysql_models.Parameter).get(body.id)
+                if parameter:
+                    parameter.name = body.name
+                    parameter.description = body.description
+                    session.commit()
+                    return ResponseModel(
+                        status="ok",
+                        message="Parameter successfully updated",
+                        data=parameter,
+                        code=200
+                    )
+                else:
+                    return ResponseModel(
+                        status="error",
+                        message="Parameter not found",
+                        data=None,
+                        code=404
+                    )
         except Exception as e:
-            print("Error updating Parameter in database")
+            print(f"Error updating Parameter in database: {e}")
             return ResponseModel(
                 status="error",
                 message=str(e),
