+++ app/controllers/parameter_room_handler.py
@@ -1,10 +1,9 @@
 import app.models.models as models
 from app.models.response_models import ResponseModel
 import app.mysql.models as mysql_models
-from app.mysql.mysql import Nexus1DataBase
-
-import app.utils.vars as var
 from sqlalchemy.orm import Session
+import app.utils.vars as var
+from app.mysql.mysql import Nexus1DataBase
 
 
 class ParameterRoom_Controller:
@@ -12,23 +11,20 @@ class ParameterRoom_Controller:
     Controller for managing ParameterRoom-related operations.
 
     This class handles the creation, updating, deletion, and retrieval of ParameterRooms in the database.
-
-    
     """
 
-    def __init__(self) -> None:
+    def __init__(self, db: object) -> None:
         """
         Initializes a new instance of the ParameterRoom_Controller class.
 
-        This constructor does not take any parameters and does not perform any operation.
+        Parameters:
+            db (object): An instance of a database (e.g., Nexus1DataBase) that will be used to create the database connection.
         """
-        pass
+        self.db = db  # Recibe la instancia de la base de datos desde fuera
     
     def healthz(self):
         """
         Checks the status of the connection.
-
-        This method returns a "ok" status message indicating that the API is working correctly.
 
         Returns:
             dict: A dictionary with the status "ok".
@@ -39,9 +35,6 @@ class ParameterRoom_Controller:
         """
         Creates a new ParameterRoom in the database.
 
-        This method takes a ParameterRoom object, which contains the necessary data to create a role
-        and saves it to the database.
-
         Parameters:
             body (models.ParameterRoom): An object containing the shelter data to create.
 
@@ -49,16 +42,15 @@ class ParameterRoom_Controller:
             ResponseModel: A response model with the status of the operation, message, and response data.
         """
         try:
-            body_row = mysql_models.ParameterRoom(id_room = body.id_room, id_parameter= body.id_parameter, date = body.date, value = body.value)
-            db = Nexus1DataBase(var.MYSQL_URL)
-            with Session(db.engine) as session:
+            body_row = mysql_models.ParameterRoom(id_room=body.id_room, id_parameter=body.id_parameter, date=body.date, value=body.value)
+            with Session(self.db.engine) as session:
                 session.add(body_row)
                 session.commit()
-                session.close()
+                session.refresh(body_row)
             return ResponseModel(
                 status="ok",
                 message="ParameterRoom inserted into database successfully",
-                data=None,
+                data=body_row,
                 code=201
             )
         except Exception as e:
@@ -74,25 +66,21 @@ class ParameterRoom_Controller:
         """
         Retrieves all ParameterRooms from the database.
 
-        This method queries all ParameterRooms records in the database and returns them.
-
         Returns:
             ResponseModel: A response model with the status of the operation, message, and ParameterRoom data.
         """
         try:
-            db = Nexus1DataBase(var.MYSQL_URL)
             response: list = []
-            with Session(db.engine) as session:
+            with Session(self.db.engine) as session:
                 response = session.query(mysql_models.ParameterRoom).all()
-                session.close()
             return ResponseModel(
                 status="ok",
-                message="All ParametersRooms successfully retrieved",
+                message="All ParameterRooms successfully retrieved",
                 data=response,
-                code=201
+                code=200
             )
         except Exception as e:
-            print("Error retrieving ParametersRooms from database")
+            print("Error retrieving ParameterRooms from database")
             return ResponseModel(
                 status="error",
                 message=str(e),
@@ -104,27 +92,31 @@ class ParameterRoom_Controller:
         """
         Deletes a ParameterRoom from the database.
 
-        This method takes a ParameterRooms object, which contains the ID of the ParameterRooms to delete.
-
         Parameters:
-            body (models.ParametersRoomDelete): An object containing the shelter ID to delete.
+            body (models.ParameterRoomDelete): An object containing the shelter ID to delete.
 
         Returns:
-            ResponseModel: A response model with the status of the operation, message, and deleted ParametersRoom data.
+            ResponseModel: A response model with the status of the operation, message, and deleted ParameterRoom data.
         """
         try:
-            db = Nexus1DataBase(var.MYSQL_URL)
-            with Session(db.engine) as session:
+            with Session(self.db.engine) as session:
                 parameterRoom_deleted = session.query(mysql_models.ParameterRoom).get(body.id)
-                session.delete(parameterRoom_deleted)
-                session.commit()
-                session.close()
-            return ResponseModel(
-                status="ok",
-                message="ParameterRoom successfully deleted",
-                data=parameterRoom_deleted,
-                code=201
-            )
+                if parameterRoom_deleted:
+                    session.delete(parameterRoom_deleted)
+                    session.commit()
+                    return ResponseModel(
+                        status="ok",
+                        message="ParameterRoom successfully deleted",
+                        data=parameterRoom_deleted,
+                        code=200
+                    )
+                else:
+                    return ResponseModel(
+                        status="error",
+                        message="ParameterRoom not found",
+                        data=None,
+                        code=404
+                    )
         except Exception as e:
             print("Error deleting ParameterRoom from database")
             return ResponseModel(
@@ -138,32 +130,34 @@ class ParameterRoom_Controller:
         """
         Updates an existing ParameterRoom in the database.
 
-        This method takes a ParameterRoom object, which contains the updated data for the ParameterRoom,
-        and updates the corresponding record in the database.
-
         Parameters:
-            body (models.ParametersRoomUpdate): An object containing the updated ParametersRoom data.
+            body (models.ParameterRoomUpdate): An object containing the updated ParameterRoom data.
 
         Returns:
-            ResponseModel: A response model with the status of the operation, message, and updated ParameterRoom Role.
+            ResponseModel: A response model with the status of the operation, message, and updated ParameterRoom data.
         """
         try:
-            db = Nexus1DataBase(var.MYSQL_URL)
-            with Session(db.engine) as session:
-                ParameterRoom: mysql_models.ParameterRoom = session.query(mysql_models.ParameterRoom).get(body.id)
-                ParameterRoom.id_room = body.id_room
-                ParameterRoom.id_parameter = body.id_parameter
-                ParameterRoom.iddate = body.date
-                ParameterRoom.idvalue = body.value
-                session.dirty  # This seems redundant; the session will be dirty when an attribute is modified
-                session.commit()
-                session.close()
-            return ResponseModel(
-                status="ok",
-                message="ParameterRoom successfully updated",
-                data=ParameterRoom,
-                code=201
-            )
+            with Session(self.db.engine) as session:
+                parameterRoom: mysql_models.ParameterRoom = session.query(mysql_models.ParameterRoom).get(body.id)
+                if parameterRoom:
+                    parameterRoom.id_room = body.id_room
+                    parameterRoom.id_parameter = body.id_parameter
+                    parameterRoom.date = body.date
+                    parameterRoom.value = body.value
+                    session.commit()
+                    return ResponseModel(
+                        status="ok",
+                        message="ParameterRoom successfully updated",
+                        data=parameterRoom,
+                        code=200
+                    )
+                else:
+                    return ResponseModel(
+                        status="error",
+                        message="ParameterRoom not found",
+                        data=None,
+                        code=404
+                    )
         except Exception as e:
             print("Error updating ParameterRoom in database")
             return ResponseModel(
